Removes complicated Clear Format logic and switches the hook to a pure function
