+++ packages/front-end/services/sqlFormatter.ts
@@ -1,0 +1,81 @@
+import { DataSourceType } from "back-end/types/datasource";
+import { format } from "shared/sql";
+import { FormatDialect } from "shared/src/types";
+
+function getSqlDialect(
+  datasourceType: DataSourceType
+): FormatDialect | undefined {
+  const typeMap: Record<DataSourceType, FormatDialect | ""> = {
+    redshift: "redshift",
+    snowflake: "snowflake",
+    mysql: "mysql",
+    bigquery: "bigquery",
+    postgres: "postgresql",
+    mssql: "tsql",
+    clickhouse: "sql",
+    growthbook_clickhouse: "sql",
+    athena: "trino",
+    presto: "trino",
+    databricks: "sql",
+    vertica: "postgresql",
+    mixpanel: "", // no formatting for mixpanel
+    google_analytics: "bigquery",
+  };
+
+  return typeMap[datasourceType];
+}
+
+// The formatter doesn't support template variables, so we need to replace them with placeholders
+function replaceTemplateVariables(
+  sql: string
+): { sql: string; placeholders: string[] } {
+  const templateRegex = /{{[^}]+}}/g;
+  const placeholders: string[] = [];
+  const sqlWithoutTemplates = sql.replace(templateRegex, (match) => {
+    placeholders.push(match);
+    return `__TEMPLATE_${placeholders.length - 1}__`;
+  });
+  return { sql: sqlWithoutTemplates, placeholders };
+}
+
+// Restore the template variables after formatting
+function restoreTemplateVariables(sql: string, placeholders: string[]): string {
+  return sql.replace(
+    /__TEMPLATE_(\d+)__/g,
+    (_, index) => placeholders[parseInt(index)]
+  );
+}
+
+export function formatSql(
+  sql: string,
+  datasourceType?: DataSourceType
+): { formattedSql: string | null; error: string | null } {
+  if (!datasourceType) {
+    return { formattedSql: null, error: "No datasource type provided" };
+  }
+
+  const dialect = getSqlDialect(datasourceType);
+  if (!dialect) {
+    return {
+      formattedSql: null,
+      error: "Formatting not supported for this datasource type",
+    };
+  }
+
+  // Format the SQL - using shared format function
+  const { sql: sqlWithoutTemplates, placeholders } = replaceTemplateVariables(
+    sql
+  );
+
+  let formatError: string | null = null;
+  const formatted = format(sqlWithoutTemplates, dialect, ({ error }) => {
+    formatError = error.message;
+  });
+
+  if (formatError || !formatted) {
+    return { formattedSql: null, error: formatError || "Failed to format SQL" };
+  }
+
+  const result = restoreTemplateVariables(formatted, placeholders);
+  return { formattedSql: result, error: null };
+}
